+++ ui/packages/design-system/src/components/Feedback/KsDialog.vue
@@ -66,6 +66,7 @@
     .kel-dialog {
         --kel-dialog-bg-color: var(--ks-bg-elevated);
         border: 1px solid var(--ks-border-default);
+        border-radius: 14px;
 
         .kel-dialog__header {
             font-size: var(--ks-font-size-base);
@@ -97,7 +98,9 @@
             padding-right: var(--kel-dialog-padding-primary);
             width: calc(100% + var(--kel-dialog-padding-primary) * 2);
             background-color: var(--ks-bg-base);
-            border-radius: 0 0  var(--kel-dialog-border-radius) var(--kel-dialog-border-radius);
+            border-bottom-left-radius: 0.875rem;
+            border-bottom-right-radius: 0.875rem;
+            overflow: hidden;
         }
     }
 </style>
